requirements.txt
python-telegram-bot~=20.6
PyYAML~=6.0.1